+++ Sources/MapboxMaps/Foundation/Events/Payload/NoPayload.swift
@@ -3,4 +3,3 @@ import Foundation
 public struct NoPayload: Decodable {
 
 }
-
