+++ Tests/MapboxMapsTests/Foundation/Mocks/MockMapboxMap.swift
@@ -65,7 +65,7 @@ final class MockMapboxMap: MapboxMapProtocol {
     let onEveryStub = Stub<OnEveryParams, Cancelable>(defaultReturnValue: MockCancelable())
     @discardableResult
     func onTypedEvery<Payload: Decodable>(_ eventType: MapEvents.Event<Payload>, handler: @escaping (TypedEvent<Payload>) -> Void) -> Cancelable {
-        onEveryStub.call(with: OnEveryParams(eventName: eventType.name, handler: { handler($0 as! TypedEvent<Payload>)} ))
+        onEveryStub.call(with: OnEveryParams(eventName: eventType.name, handler: { handler($0 as! TypedEvent<Payload>)}))
     }
 
     let beginAnimationStub = Stub<Void, Void>()
