+++ Tests/MapboxMapsTests/Snapshot/SnapshotterTests.swift
@@ -93,6 +93,7 @@ final class SnapshotterTests: XCTestCase {
             XCTAssertIdentical(handlerStub.invocations.first?.parameters, event)
         }
 
+        // swiftlint:disable opening_brace
         let eventInvocations = [
             { try verifyInvocation(eventType: .mapLoaded) },
             { try verifyInvocation(eventType: .mapLoadingError) },
@@ -109,6 +110,8 @@ final class SnapshotterTests: XCTestCase {
             { try verifyInvocation(eventType: .cameraChanged) },
             { try verifyInvocation(eventType: .resourceRequest) }
         ]
+        // swiftlint:enable opening_brace
+
         try eventInvocations.randomElement()!()
     }
 
@@ -153,6 +156,7 @@ final class SnapshotterTests: XCTestCase {
             XCTAssertIdentical(handlerStub.invocations.first?.parameters, event)
         }
 
+        // swiftlint:disable opening_brace
         let eventInvocations = [
             { try verifyInvocation(eventType: .mapLoaded) },
             { try verifyInvocation(eventType: .mapLoadingError) },
@@ -169,6 +173,8 @@ final class SnapshotterTests: XCTestCase {
             { try verifyInvocation(eventType: .cameraChanged) },
             { try verifyInvocation(eventType: .resourceRequest) }
         ]
+        // swiftlint:enable opening_brace
+
         try eventInvocations.randomElement()!()
     }
 }
